Add initial stored procedures
They can fetch the total score that you have based on steamid, and for
which mode you are running. The idea is that the plugin will use the
servers default mode for calculating score.

The procedures also only look at maps that you have added into the local
mappool with the Local Ranks mappool file and update command. This
solves the issue of having to somehow sync the mapcycle file into SQL.