Initial setup of DB table
Unfortunately I cannot figure out if it would be possible to create the
stored procedures from the plugin, so that will have to wait until
later. Maybe it is not the end of the world if they have to be created
manually? Although it sucks a bit considering a lot of the logic is in
them, and updates will be tedious...