Prevent players from getting duplicate points from map
Instead of getting all their runs (which is wrong, since you have many
runs per map in GOKZ DB), we are just getting a list of all the
MapCourseIds that the player has finished. Then I changed the
GetMapScore procedure to not take TPs as an argument anymore, but
instead always calculates both TP/PRO and PRO ranks and score.