Add SP and command to recalculate top scores
Stored procedure that can update the score for X amount of top players,
and a command that runs it for the game modes, on the top 100 players.

This is useful since scores are not updated if players are not joining
the server. So you can use this command to clean out those players that
have lost their scores due to other places taking their rank.