Add skillgroups and !ranks command
Possibility to add skillgroups into a config file, which will be read,
and then you can output them all with the !ranks command. Currently
nothing else is using them, and I am really not sure if the
implementation is a really smart one. That we will see later when we get
to the stage of trying to actually use the skill groups in some places..