Add forward for initial score change
Calls when the player first joins the server and the score is calculated
for the player. Can be used for things that should be checked only when
the player joins the first time, like connection or rank announcements,
or in this case, for kicking players that are not high enough rank.