Add Kicker plugin
Allows you to set Score or Skillgroup limits for your server.

The score limit will require players to reach a certain score to be able
to play on the server, else they will get kicked.

The skillgroup limit will check that players are a high enough
skillgroup to be allowed to play.