Correctly count rank for players with 0 score.
If a player has 0 score, the rank has been calculated as the total
amount of players in the DB, as everyone had >= 0 score. So now instead
we are calculating the amount of players that have more score than you,
and then adding +1 to get the ranks to start from 1.

This means though that if you have 0 score, your rank will be total
amount of ranked players +1, which will be bigger than the total amount
of ranked players. That is why we need to set the lowest rank to be the
same as the total rank. So essentially, if you do not have any score,
you will be ranked (total ranked players + 1)/(total ranked players + 1).