Change formatting of chat messages
Some changes to the formatting of chat messages:

1. The [ is now not before the skillgroup, but before the mode, so the
chat will now look like `[KZT | Skilled] Playername : Some text...`.
This makes it easier to "ignore" that part of the message, as you can
just ignore everything inside [] when you read the message.

2. The mode will now have a purple color. This color is used in other
messages also from GOKZ, so it makes sense to also use it here. It
further makes it easier to read the tag, as the [] and | will be default
(whiteisch) color, so there is more contrast with the Mode.

3. For spectators, the * that shows you are spectating has been moved
from before the mode/skillgroup, to after. This means that all chat will
start off the same way, which is clearer. You can also more easily
ignore the [] part of the chat, and start with reading from * playername
: message, so there is perhaps less for you to parse in the chat
messages when reading.