Use PRO rank for calculating top20 bonus points
Bad copypasta made the PRO score bonus to be calculated using your TP
rank. So the scores for players might be off quite a bit in some cases.
Now instead we use the PRO rank to calculate the PRO top20 bonus, and
the TP rank as normal for calculating the TP rank.