Change delimeter from ; to // for drop procedure call
Since we are changing the delimeter already before, we should use that
delimeter to run the drop statement before going on with the create
statement.

The ; worked for me, since I copy-pasted this content into mysql, so
even if there was an invalid delimeter, it still somehow executed the
queries properly it seems. However, when importing this into MySQL DBs
through PHPMyAdmin or similar, there have been problems with getting
these files to run. This change seems to make it work fine in
PHPMyAdmin.