Create initial tags after admin check instead of when joining
The OnClientPostAdminCheck always fires after the OnClientPutInServer,
so there should not be much difference. This way though, the admin check
has been done. Without this change, it has been observed that sometimes
players can join the server without getting their ADMIN or VIP tag in
scoreboard, so hopefully this will fix it.