Skip score calculation if map not in ranked pool
Shows a message instead that says that the map is not in the ranked
pool, and that points will be awarded later. This will (hopefully) avoid
some confusion if new maps are added to server, and pool is not updated.
And remove some negative comments such as "good shit, finished with good
time and got 0 score", as without this fix, that would be the case.