Set clantag when GOKZ options are loaded
Previously we only set the clantag when the client connected (which was
later changed to after admin check). However, there is no guarantee that
the clients cookies, and in extension, the GOKZ options, have loaded at
that point. This has lead to the scoreboard reporting the wrong mode for
players, often even the whole server as VNL. Hopefully this addition
will now make it so that the correct mode is always set, both for VIP
and ADMINs, but also for all other players.

The reason this has not been seen that often is probably because GOKZ
has relatively quick queries, so all cookies are probably loaded at the
time the player joins, and therefor he gets the correct mode. Now
however, we are also running the Myranks calculations on the same DB
connection, so on mapchange with many players, the VNL rank bug appears.