Calculate score with low priority in initial setup
We first run a transaction for the user with high prio that makes sure
that the user exists in the db, then we queue up another one later on
with low prio, that will do the heavy lifting and calculate score.

The idea with this is to allow for the important gokz (and myranks)
stuff load first, before we move on to calculating the score. Getting
the score calculated is not very important anyways, and the only reason
I didnt do it like this to begin with, was to avoid the extra
transactions.