Wait for score and admin rights to load before kick
Without this change, the client is checked for kick once the score is
loaded in the myrank plugin, which happens after client authorization.
There is nothing that checks that the admin rights are loaded though,
which means that sometimes people that should be allowed on the server
get kicked anyways.

This change keeps track of the clients that connect, and if their score
and admin rights have been loaded. When both are loaded, then it will
proceed with actually checking if the player should be kicked.